Remove unnecessary level copies
Each level was being memberwise completely copied in sokoban::move_ to
local variables that were then destroyed. This was vestigial from when
levels were created directly from the mutated copies of the data
members. This is even faster than the shared_ptr tiles design, but
doesn't improve the memory usage.